+++ src/types/point.rs
@@ -18,7 +18,7 @@ use crate::{FromTokens, Geometry, WktNum};
 use std::fmt;
 use std::str::FromStr;
 
-#[derive(Clone, Debug, Default)]
+#[derive(Clone, Debug, Default, PartialEq)]
 pub struct Point<T: WktNum>(pub Option<Coord<T>>);
 
 impl<T> Point<T>